+++ v2.0/pic/PIC18F2431/c/dual.seq.v1.0.X/src/leds.h
@@ -4,7 +4,7 @@
 void leds_init(void);
 void leds_blink(void);
 
-void leds_set_mode(unsigned int value);
-void leds_set_target(unsigned int value);
+void leds_set_mode(unsigned short value);
+void leds_set_target(unsigned short value);
 
 #endif
